new rule: too many downvotes per user and day #5
I needed a way to count downvotes per user and day. We have a 'vote log' for things like this, but there was no way to specify 'downvotes' or filter, because the log treats upvotes and downvotes the same way.
To avoid conflict with previous versions of the plugin I have added a new log (user:day:type) that allows us to count downvotes (and upvotes) per user and day.