when you undo a vote, you don't know vote.type
and if you don't know vote.type, you cannot remove the vote from the log.
and if the log is not deleted that's a problem.
to fix it we now delete the log from both upvote and downvote logs. Yes, the vote will only be in one of those two logs, but we can try and delete it from both sets without problems. Not optimal, but simple.